+++ dosing-device/src/main/java/app/web/adapter/persistence/DosingSuperEntity.java
@@ -23,5 +23,5 @@ class DosingSuperEntity {
     protected int amountCorrectMeasures;
     protected int amountAboveMeasures;
     protected float averageMeasure;
-    protected float correctMeasurePercent;
+    protected int correctMeasurePercent;
 }
